// resources/js/game/reusable-components/item/enums/item-details-section-labels.ts
export enum ItemDetailsSectionLabels {
  COST = 'Cost',
  CRAFTING_REQ = 'Crafting (Req.)',
  CRAFTING_TRIVIAL = 'Crafting (Trivial)',
  DAMAGE = 'Damage',
  BONUS_DAMAGE_MOD = 'Bonus Damage Mod %',
  BONUS_DAMAGE = 'Bonus Damage %',
  COUNTER = 'Counter',
  AMBUSH = 'Ambush',
  COUNTER_RESISTANCE = 'Counter resistance',
  AMBUSH_RESISTANCE = 'Ambush Resistance',
  AC = 'AC',
}